requirements.txt
# Core Streamlit and WebRTC
streamlit>=1.29.0
streamlit-webrtc>=0.47.1
typing-extensions>=4.0.0

# Computer Vision and AI
mediapipe>=0.10.13
google-generativeai>=0.3.2
openai-whisper>=20231117
numpy>=1.24.0
opencv-python-headless>=4.8.0
pillow>=10.0.0

# Audio processing (optional - Whisper dependencies)
torch>=2.0.0
torchvision>=0.15.0

# Testing
hypothesis>=6.88.0
pytest>=7.4.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0